requirements.txt
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
python-dotenv==1.0.1
loguru==0.7.2
dvc==3.50.2
pytest==8.3.3
apache-airflow==2.9.2
